+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Fix bug on quoted attribute parser.
+- Remove trailing line-ending from source line of syntax error message.
 
 ## 0.4.1 - 2021-12-26